… packages/api/package.json
… packages/beacon-node/package.json
… packages/beacon-node/src/chain/opPools/opPool.ts
… packages/beacon-node/src/chain/options.ts
… packages/beacon-node/src/chain/prepareNextSlot.ts
… packages/beacon-node/src/chain/regen/queued.ts
… packages/beacon-node/src/chain/regen/regen.ts
… packages/beacon-node/src/chain/shufflingCache.ts
… packages/beacon-node/src/chain/stateCache/index.ts
… packages/beacon-node/src/chain/stateCache/lruBlockStateCache.ts
… packages/beacon-node/src/chain/stateCache/memoryCheckpointsCache.ts
… packages/beacon-node/src/chain/stateCache/persistent/db.ts
… packages/beacon-node/src/chain/stateCache/persistent/file.ts
… packages/beacon-node/src/chain/stateCache/persistent/types.ts
… packages/beacon-node/src/chain/stateCache/persistentCheckpointsCache.ts
… packages/beacon-node/src/chain/stateCache/stateContextCache.ts
… packages/beacon-node/src/chain/stateCache/stateContextCheckpointsCache.ts
… packages/beacon-node/src/chain/stateCache/types.ts
… packages/beacon-node/src/chain/validation/aggregateAndProof.ts
… packages/beacon-node/src/chain/validation/attestation.ts
… packages/beacon-node/src/chain/validation/signatureSets/aggregateAndProof.ts
… packages/beacon-node/src/chain/validation/signatureSets/selectionProof.ts
… packages/beacon-node/src/db/beacon.ts
… packages/beacon-node/src/db/buckets.ts
… packages/beacon-node/src/db/interface.ts
… packages/beacon-node/src/db/repositories/checkpointState.ts
… packages/beacon-node/src/metrics/metrics/lodestar.ts
… packages/beacon-node/src/util/array.ts
… packages/beacon-node/src/util/multifork.ts
… packages/beacon-node/src/util/sszBytes.ts