**Motivation**

- It's not sustainable to add more states to our cache, lodestar is not able to be alive in long unfinality period
- Also when memory is high, lodestar has bad performance due to a lot of `gc` calls (specifically state transition function)
- We're now able to restore a state from bytes in a way that there is only 1 base state across lodestar

**Description**

- New `loadState` api to load state bytes into an existing TreeBacked state
  - Since it mostly share the validators array, memory only increase ~100MB
  - Time to load state and do the hashTreeRoot is from 2s to 3s on with mainnet state
- Checkpoint state cache
  - Only store up to `maxEpochsInMemory` (default = 2) epochs. If more than that persist the states to disk. Right now we persist 2 checkpoint states (spec + non-spec) to disk per epoch
  - Maintain an order of states to choose good state to prune from memory (and persist to disk)
    - Persist to `fs` through `FilePersistentApis` or `db` through `DbPersistentApis`, cleanup at startup
    - by default, persist to `db`
  - The cache may contain a `CachedBeaconState` or a persistent key
  - Find closest state and reload from disk in "getOrReload" apis based on persistent key
  - Also maintain "get" apis
  - Maintain the old implementation through `MemoryCheckpointStateCache` and `CheckpointStateCache` interface
- State cache
  - Store maxStates (default = 32, but 4 worked in Holesky) consistently
  - Maintain order of state to make it's easier to select a state to prune. Head state always stay on top
  - Maintain the old implementation through `StateContextCache` and `BlockStateCache` interface

- Regen
  - With reload capability, we can regen from whatever checkpoint state. Reload takes 2s - 3s, and rarely happen
    - For important flows, like "getPreState" or "updateHeadState", use "getOrReload" apis which may reload checkpoint states from disk
    - For other flows, use the normal "get" apis 
- Feature flag:
  - `nHistoricalStates` flag to use the new `PersistentCheckpointStateCache` and `LRUBlockStateCache`
  - by default, go with the old implementation - `MemoryCheckpointStateCache` and `StateContextCache` 
- Shuffling cache
  - Populate cache when we cross through epoch boundary
  - Used when we reloading a state from disk, or attestation verification
  - This is technically not part of this PR but a prerequisite, will make a separate PR for this

**Test result on Holesky**
- This branch was able to survive under non-finality time of holesky with default heap size config
  - Node 1 (maxEpochsInMemory=2 & maxStates=32): no reload happened
  - Node 2 (maxEpochsInMemory=0 & maxStates=4): reloaded > 15 times, average reload duration = 2s (this did not include the hashTreeRoot)
  
- Used heap is reduced based on `maxEpochsInMemory` (for checkpoint cache) and `maxStates` (for state cache)

|tested node|used heap (GB)|
|------------|----------|
|regular holesky (when stable)| 4.7 to 5.7|
|2 epochs - 32 maxStates|3.2 to 3.8|
|1 epoch - 16 maxStates|2.95 to 3.7|
|0 epoch - 4 maxStates|2.8 to 3.4|

**Some down sides with this approach**
- We are not able to access the finalized tree backed state, we can only get finalized state bytes from file
- We have to store 2 checkpoint states per epoch which take 750ms * 2 = 1.5s. It choose to persist to disk at the last 1/3 slot of the 1st slot of epoch but this could be improved to store only the spec checkpoints

part of #5968

**TODOs**

- [x] Use shuffling cache for attestation verification
- [x] At each epoch we store both spec checkpoint and non-spec checkpoint and persist both of them. May want to only persist spec checkpoint (it takes 750ms to 900ms to serialize 1 Holesky state)
- [x] Investigate this issue: With default config of maxEpochsInMemory=2 and maxStates=32, heap size is a little bit smaller but rss increased more compared to unstable => this is due to `subscribeAllSubnets` flag
- [x] Find the best default config => based on used heap, `2 epochs - 32 maxStates` seems to be the best default config
- [x] Confirm same state type before reloading state => modified `loadState` to bypass this check
- [x] Investigate high state transition time and memory issue on beta mainnet node => this is due to subscribeAllSubnets flag, and partially ssz
- [x] Refactor to have a feature flag to turn this feature on/off as it may take weeks to test
- [x] Track epochs in memory and epochs in disk
- [ ] Test and fix bugs